+++ LuaScriptBuilder.php
@@ -215,7 +215,7 @@ EOF;
             return $str;
         }
 
-        return "'" . addcslashes(str_replace("'", "\\'", $str), "\000\n\r\\\032") . "'";
+        return "'" . addcslashes($str, "\000\n\r\\\032\047") . "'";
     }
 
     /**