+++ CHANGELOG.md
@@ -4,8 +4,8 @@ Yii Framework 2 redis extension Change Log
 2.0.5 under development
 -----------------------
 
+- Bug #37: Fixed detection of open socket (mirocow)
 - Chg #14: Added missing `BLPOP` to `$redisCommands` (samdark)   
-
 
 2.0.4 May 10, 2015
 ------------------