+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 redis extension Change Log
 2.0.5 under development
 -----------------------
 
+- Bug #22: Fixed string escaping issue in LuaScriptBuilder (vistart)
 - Bug #37: Fixed detection of open socket (mirocow)
 - Bug #46: Fixed bug to execute session_regenerate_id in PHP 7.0 (githubjeka)
 - Chg #14: Added missing `BLPOP` to `$redisCommands` (samdark)